Move static helper methods to their belonging object
